refactors auth to use context internally
