+++ src/authentication/Authenticator.js
@@ -1,0 +1,12 @@
+import React, { useContext } from 'react';
+import AuthContext from './context';
+
+function Authenticator(props) {
+  const { children } = props;
+
+  return children({
+    ...useContext(AuthContext)
+  });
+}
+
+export default Authenticator;
