+++ src/authentication/context.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const AuthContext = React.createContext();
+
+export default AuthContext;
