+++ src/index.js
@@ -1,9 +1,9 @@
-import { AuthContext, AuthProvider } from './authentication';
+import { AuthProvider, Authenticator } from './authentication';
 import { FirestoreProvider, Collection, Document } from './firestore';
 
 export {
-  AuthContext,
   AuthProvider,
+  Authenticator,
   FirestoreProvider,
   Collection,
   Document
